Solving PR Comments, New Structure
