--- public/css/reset.css
@@ -1,9 +1,0 @@
-
-  *{
-	margin: 0;
-	padding: 0;
-	list-style: none;
-	text-decoration: none;
-	border: none;
-	outline: none;
-}
