+++ public/index.html
@@ -5,7 +5,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@100;300;500;900&display=swap" rel="stylesheet">
-    <link href="./css/reset.css" rel="stylesheet" />
     <title>Netflix Roulette</title>
 </head>
 <body>
