+++ src/design/settings/colors.scss
@@ -1,0 +1,7 @@
+$white-color: #FFFFFF;
+$black-color: #000000;
+$red-color: #F65261;
+$ultra-light-grey-color: #979797;
+$light-grey-color: #555555;
+$grey-color: #424242;
+$darkGrey-color: #232323;
