+++ src/design/settings/fonts/fonts.scss
@@ -1,0 +1,32 @@
+@import "./font-families.scss";
+
+@font-face {
+  font-style: normal;
+  font-weight: 300;
+  src: local('Montserrat Light'), local('Montserrat-Light'), url(https://fonts.gstatic.com/s/montserrat/v14/JTURjIg1_i6t8kCHKm45_cJD3gnD_g.woff2) format('woff2');
+  unicode-range: U+0000-00FF, U+0131, U+0152-0153, U+02BB-02BC, U+02C6, U+02DA, U+02DC, U+2000-206F, U+2074, U+20AC, U+2122, U+2191, U+2193, U+2212, U+2215, U+FEFF, U+FFFD;
+}
+
+.xs{
+  font-size: 14px;
+}
+
+.s{
+  font-size: 16px;
+}
+
+.m{
+  font-size: 18px;
+}
+
+.l{
+  font-size: 20px;
+}
+
+.xl{
+  font-size: 24px;
+}
+
+.xxl{
+  font-size: 40px;
+}
