+++ src/pages/home/home.scss
@@ -1,4 +1,4 @@
-@import "../../patterns/theme/common.scss";
+@import "common.scss";
 
 body{
   display: grid;
