--- src/patterns/atoms/MovieTitle/MovieTitle.scss
@@ -1,9 +1,0 @@
-@import "../../theme/common.scss";
-
-.movieTitle{
-  color: var(--c-white);
-  font-size: 18px;
-  font-weight: 300;
-  margin-bottom: 5px;
-  opacity: 70%;
-}
