--- src/patterns/atoms/Text/Text.scss
@@ -1,7 +1,0 @@
-@import "../../theme/common.scss";
-
-.text{
-  color: var(--c-white);
-  font-weight: 300;
-  opacity: 50%;
-}
