--- src/patterns/atoms/Title/Title.scss
@@ -1,6 +1,0 @@
-@import "../../theme/common.scss";
-
-.title{
-  color: var(--c-white);
-  font-weight: 100;
-}
