--- src/patterns/molecules/MainBanner/MainBanner.scss
@@ -1,50 +1,0 @@
-@import "../../theme/common.scss";
-
-.MainBanner{
-    align-items: center;
-    background-image: url(../../../assets/images/banner.png);
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: column;
-    flex-wrap: wrap;
-    height: 396px;
-    padding: 20px 60px;
-    form{
-        display: flex;
-        width: 100%;
-        input{
-            width: 80%;
-        }
-        button{
-            background-color: var(--c-red);
-            border-radius: 5px;
-            color: var(--c-white);
-            font-family: $montserrat;
-            font-size: 20px;
-            width: 20%;
-        }
-    }
-    section{
-        display: flex;
-        justify-content: space-between;
-        width: 100%;
-        p {
-            color: var(--c-red);
-            font-size: 20px;
-            font-weight: 900;
-            span {
-                font-weight: 400;
-            }
-        }
-    }
-    .inputMain {
-        background: #323232;
-        border-radius: 5px;
-        box-sizing: border-box;
-        color: #FFF;
-        font-size: 20px;
-        height: 57px;
-        opacity: 80%;
-        padding: 20px;
-    }
-}
