--- src/patterns/molecules/MovieElement/MovieElement.scss
@@ -1,10 +1,0 @@
-@import "../../theme/common.scss";
-
-.movieElement{
-    &-info{
-        box-sizing: border-box;
-        display: flex;
-        justify-content: space-between;
-        padding: 15px 0;
-    }
-}
