+++ src/ui-kit/atoms/MovieTitle/MovieTitle.scss
@@ -1,0 +1,9 @@
+@import "common.scss";
+
+.movieTitle{
+  color: var(--c-white);
+  font-size: 18px;
+  font-weight: 300;
+  margin-bottom: 5px;
+  opacity: 70%;
+}
