+++ src/ui-kit/atoms/Text/Text.scss
@@ -1,0 +1,7 @@
+@import "common.scss";
+
+.text{
+  color: var(--c-white);
+  font-weight: 300;
+  opacity: 50%;
+}
