+++ src/ui-kit/atoms/Title/Title.scss
@@ -1,0 +1,6 @@
+@import "common.scss";
+
+.title{
+  color: var(--c-white);
+  font-weight: 100;
+}
