--- tsconfig.ts
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "esModuleInterop": true,
-        "target": "es2017",
-        "skipLibCheck": true,
-        "noImplicitAny": true,
-        "noEmit": true,
-        "moduleResolution": "node",
-        "sourceMap": true,
-        "outDir": "dist",
-        "baseUrl": ".",
-        "jsx": "react",
-        "typeRoots" : ["@types"]
-    },
-
-}
